Increase psalm to level 4 (#2016)
* Increase psalm to level 4

* Fix phpstan-* annotations with its normal annotation

* change private property to private const

+++ src/Resizer/SquareResizer.php
@@ -73,7 +73,7 @@ final class SquareResizer implements ResizerInterface
             $crop = $higher - $lower;
 
             if ($crop > 0) {
-                $point = $higher === $size->getHeight() ? new Point(0, 0) : new Point($crop / 2, 0);
+                $point = $higher === $size->getHeight() ? new Point(0, 0) : new Point((int) ($crop / 2), 0);
                 $image->crop($point, new Box($lower, $lower));
                 $size = $image->getSize();
             }
